Check for valid daemon in TearDownTest.

+++ integration-cli/check_test.go
@@ -302,14 +302,16 @@ func (s *DockerSwarmSuite) TearDownTest(c *check.C) {
 	testRequires(c, DaemonIsLinux)
 	s.daemonsLock.Lock()
 	for _, d := range s.daemons {
-		d.Stop()
-		// raft state file is quite big (64MB) so remove it after every test
-		walDir := filepath.Join(d.root, "swarm/raft/wal")
-		if err := os.RemoveAll(walDir); err != nil {
-			c.Logf("error removing %v: %v", walDir, err)
-		}
+		if d != nil {
+			d.Stop()
+			// raft state file is quite big (64MB) so remove it after every test
+			walDir := filepath.Join(d.root, "swarm/raft/wal")
+			if err := os.RemoveAll(walDir); err != nil {
+				c.Logf("error removing %v: %v", walDir, err)
+			}
 
-		cleanupExecRoot(c, d.execRoot)
+			cleanupExecRoot(c, d.execRoot)
+		}
 	}
 	s.daemons = nil
 	s.daemonsLock.Unlock()
